Better design of the `CodeAndLocationProvider`
Instead of directly setting the code and location provider in the implementation of the `CodeAndLocationProvider` we rather now have an implementor of this interface implement the methods `codeOf` and `locationOf` and set the code/location in the node builder. This is more in line with the other providers, since they should *provide* a value, but not *set* it.

The function `setCodeAndLocation` will stay for now for legacy reason and is marked deprecated. We seem to have a lot of language frontends that directly call this function, which is bad design anyway. One should supply the raw node to the node builder instead.
